# specs/003-auth-multi-user/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo API with Authentication
  description: API for managing todo items with user authentication and authorization
  version: 1.0.0
servers:
  - url: http://localhost:7861
    description: Development server

paths:
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: User already exists or invalid input

  /auth/login:
    post:
      summary: Authenticate user and get JWT token
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: User's email
                password:
                  type: string
                  description: User's password
              required:
                - username
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: JWT token
                  token_type:
                    type: string
                    description: Token type (usually "bearer")
                  user:
                    $ref: '#/components/schemas/UserResponse'
        '401':
          description: Invalid credentials

  /auth/me:
    get:
      summary: Get current authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized

  /todos:
    get:
      summary: Get all todos for the authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TodoResponse'
        '401':
          description: Unauthorized
    post:
      summary: Create a new todo for the authenticated user
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoCreate'
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '401':
          description: Unauthorized

  /todos/{id}:
    get:
      summary: Get a specific todo by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Todo data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '401':
          description: Unauthorized
        '404':
          description: Todo not found
    put:
      summary: Update a specific todo by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoUpdate'
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '401':
          description: Unauthorized
        '404':
          description: Todo not found
    delete:
      summary: Delete a specific todo by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Todo deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Todo not found

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    UserBase:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 255
      required:
        - email

    UserCreate:
      allOf:
        - $ref: '#/components/schemas/UserBase'
        - type: object
          properties:
            password:
              type: string
              minLength: 1
          required:
            - password

    UserResponse:
      allOf:
        - $ref: '#/components/schemas/UserBase'
        - type: object
          properties:
            id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time

    TodoBase:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 1000
          nullable: true
      required:
        - title

    TodoCreate:
      allOf:
        - $ref: '#/components/schemas/TodoBase'

    TodoUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 1000
          nullable: true
        completed:
          type: boolean

    TodoResponse:
      allOf:
        - $ref: '#/components/schemas/TodoBase'
        - type: object
          properties:
            id:
              type: string
              format: uuid
            completed:
              type: boolean
              default: false
            created_at:
              type: string
              format: date-time
            user_id:
              type: string
              format: uuid
          required:
            - id
            - completed
            - created_at
            - user_id

security:
  - BearerAuth: []